fix: 공포도·난이도의 0.5 단위를 잃지 않는다 (D10)
`Int` 로 받다가 값이 조용히 망가지고 있었다.

  공포   "3.5"   → 35
  공포   "0.5"   →  5    ← "거의 안 무섭다" 가 "최고 공포" 로 뒤집힘
  난이도 size25  → 25    ← 화면에 별이 25 개 찍힘

`reservation.css` 가 size1·size15·size2·size25 … size45 를 각각 다르게
그린다. 두 자리는 반 칸이고 `size25` 는 2.5 다. 세 자리 이상은 우리가
모르는 표기라 넘겨짚지 않고 null 로 둔다.

`V2` 에서 컬럼을 DOUBLE PRECISION 으로 바꾸고 기존 값은 전부 비웠다.
저장된 5 가 진짜 5.0 인지 뒤집힌 0.5 인지 구분할 수 없다.

건대점 픽스처는 평점이 전부 정수라 이 버그를 통과시켰다.
소수가 든 대전점 픽스처를 따로 넣었다 — 표본이 하나면 그 표본에
없는 경우는 영원히 안 잡힌다.

+++ src/main/kotlin/com/my_dream/server/api/ApiDtos.kt
@@ -42,9 +42,9 @@ data class ThemeDto(
     /** [capacity] 에서 뽑아낸 값. 인원이 안 맞는 테마를 걸러내라고 있다 */
     val minPeople: Int?,
     val maxPeople: Int?,
-    val horrorLevel: Int?,
-    /** 별 개수 (1~5) */
-    val difficulty: Int?,
+    val horrorLevel: Double?,
+    /** 별 개수. **0.5 단위** 다 — `2.5` 가 실제로 온다 */
+    val difficulty: Double?,
     val posterUrl: String?,
 
     val sessions: List<SessionDto>,
